+++ tests/dataset_test.cpp
@@ -3,8 +3,8 @@
 
 TEST(DatasetTest, Constructor) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 0.23, 1.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 0.23, 1.0 };
 
   ogbt::Dataset dataset(data_categorical, data_dense, target);
 
@@ -21,11 +21,11 @@ TEST(DatasetTest, Constructor) {
 
 TEST(DatasetTest, PreprocessTest) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 1.0, 1.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 1.0, 1.0 };
   std::vector<std::vector<int>> data_categorical_test = data_categorical;
-  std::vector<std::vector<double>> data_dense_test = data_dense;
-  std::vector<double> target_test = target;
+  std::vector<std::vector<float>> data_dense_test = data_dense;
+  std::vector<float> target_test = target;
 
   ogbt::Dataset dataset(data_categorical, data_dense, target);
   EXPECT_EQ(data_dense, data_dense_test);
