+++ tests/loss_test.cpp
@@ -3,8 +3,8 @@
 
 TEST(MSE, score) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 0.23, 1.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 0.23, 1.0 };
 
   ogbt::Dataset dataset(data_categorical, data_dense, target);
 
@@ -17,8 +17,8 @@ TEST(MSE, score) {
 
 TEST(LogLoss, score) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 0.0, 0.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 0.0, 0.0 };
 
   ogbt::Dataset dataset(data_categorical, data_dense, target);
   std::cout << ogbt::LogLoss::score(target, dataset) << std::endl;
